feat(k8s): 워크스테이션 NAS 스택 추가(qBittorrent/Jellyfin/SFTP) 및 MetalLB 전환
 - nas 네임스페이스 도입 및 PVC 구성(qbittorrent config/complete/incomplete, jellyfin config/media, sftp extra-data)
 - qBittorrent 스택: NordLynx 사이드카, Vuetorrent 모드, PVC 마운트, OAuth2 Proxy 보호 Ingress, Cloudflare 레코드 연동
 - SFTP 스택: RSA 키 자동 로테이션(time/tls/local), 공개키 ConfigMap, atmoz/sftp 배포 및 공유 PVC 마운트
 - Jellyfin 스택 추가
 - 시스템 스택: NodePort 제거 → MetalLB 포트 기반 노출로 마이그레이션, files → nas 리네이밍
 - cert-manager CRD 스택 추가(OKE, 워크스테이션), MetalLB CRD 스택 추가(워크스테이션)
 - 불필요한 files.stack.ts 제거 및 인덱스/앱 스택 정리(argo-cd, ingress-controller, istio, oauth2-proxy 등)
 - Cloudflare 레코드 스택 및 .projenrc.ts 업데이트

BREAKING CHANGE: 워크스테이션 SFTP/토렌트 노출 방식이 NodePort에서 MetalLB로 변경됨. 관련 인그레스/방화벽/포트 설정 점검 필요.
