+++ src/terraform/stacks/k8s/oke/apps/argo-cd.stack.ts
@@ -176,7 +176,7 @@ export class K8S_Oke_Apps_ArgoCd_Stack extends AbstractStack {
   //   },
   // );
 
-  // private readonly metadata = this.provide(Resource, 'metadata', () => [
+  // metadata = this.provide(Resource, 'metadata', () => [
   //   {},
   //   this.k8sOkeSystemStack.applicationMetadata.shared.argoCd,
   // ]);
