+++ src/terraform/stacks/k8s/oke/apps/cloudbeaver.stack.ts
@@ -49,7 +49,7 @@ export class K8S_Oke_Apps_Cloudbeaver_Stack extends AbstractStack {
     },
   };
 
-  private readonly metadata = this.provide(Resource, 'metadata', () => [
+  metadata = this.provide(Resource, 'metadata', () => [
     {},
     this.k8sOkeSystemStack.applicationMetadata.shared.cloudbeaver,
   ]);
