+++ src/terraform/stacks/k8s/oke/apps/dashboard.stack.ts
@@ -53,7 +53,7 @@ export class K8S_Oke_Apps_Dashboard_Stack extends AbstractStack {
     },
   };
 
-  private readonly metadata = this.provide(Resource, 'metadata', () => [
+  metadata = this.provide(Resource, 'metadata', () => [
     {},
     this.k8sOkeSystemStack.applicationMetadata.shared.dashboard,
   ]);
