+++ src/terraform/stacks/k8s/oke/apps/home-l2tp-vpn-proxy.stack.ts
@@ -50,7 +50,7 @@ export class K8S_Oke_Apps_HomeL2tpVpnProxy_Stack extends AbstractStack {
     },
   };
 
-  private readonly metadata = this.provide(Resource, 'metadata', () => [
+  metadata = this.provide(Resource, 'metadata', () => [
     {},
     this.k8sOkeSystemStack.applicationMetadata.shared.homeL2tpVpnProxy,
   ]);
