+++ src/terraform/stacks/k8s/oke/apps/oauth2-proxy.stack.ts
@@ -59,7 +59,7 @@ export class K8S_Oke_Apps_OAuth2Proxy_Stack extends AbstractStack {
     },
   };
 
-  private readonly metadata = this.provide(Resource, 'metadata', () => [
+  metadata = this.provide(Resource, 'metadata', () => [
     {},
     {
       ...this.k8sOkeSystemStack.applicationMetadata.shared.oauth2Proxy,
