+++ src/terraform/stacks/k8s/oke/apps/redis-ui.stack.ts
@@ -51,7 +51,7 @@ export class K8S_Oke_Apps_RedisUi_Stack extends AbstractStack {
     },
   };
 
-  private readonly metadata = this.provide(Resource, 'metadata', () => [
+  metadata = this.provide(Resource, 'metadata', () => [
     {},
     this.k8sOkeSystemStack.applicationMetadata.shared.redisUi,
   ]);
