+++ src/terraform/stacks/k8s/oke/apps/rollout-trigger.stack.ts
@@ -46,7 +46,7 @@ export class K8S_Oke_Apps_RolloutTrigger_Stack extends AbstractStack {
     },
   };
 
-  private readonly metadata = this.provide(Resource, 'metadata', () => [
+  metadata = this.provide(Resource, 'metadata', () => [
     {},
     this.k8sOkeSystemStack.applicationMetadata.shared.rolloutTrigger,
   ]);
