+++ src/terraform/stacks/k8s/workstation/apps/index.ts
@@ -1,7 +1,25 @@
+// dashboard
 export * from './dashboard.stack';
+
+// istio
 export * from './istio.stack';
+
+// longhorn
 export * from './longhorn.stack';
-export * from './files.stack';
+
+// nas
+export * from './nas.jellyfin.stack';
+export * from './nas.qbittorrent.stack';
+export * from './nas.sftp.stack';
+export * from './nas.stack';
+
+// metallb
 export * from './metallb.stack';
+export * from './metallb.crd.stack';
+
+// cert-manager
 export * from './cert-manager.stack';
+export * from './cert-manager.crd.stack';
+
+// ingress-controller
 export * from './ingress-controller.stack';
